Proform 2000 Pro -3% incline (Issue #2092)

+++ src/devices/proformtreadmill/proformtreadmill.cpp
@@ -1564,7 +1564,7 @@ void proformtreadmill::characteristicChanged(const QLowEnergyCharacteristic &cha
     } else {
         if (!proform_cadence_lt) {
             Inclination =
-                (double)(((int16_t)((int16_t)newValue.at(13)) << 8) + (int16_t)((uint8_t)newValue.at(12))) / 100.0;
+                (double)(((int16_t)((int8_t)newValue.at(13)) << 8) + (int16_t)((uint8_t)newValue.at(12))) / 100.0;
         }
         Speed = (double)(((uint16_t)((uint8_t)newValue.at(11)) << 8) + (uint16_t)((uint8_t)newValue.at(10))) / 100.0;
         if (watts(weight))
